tasks 1-14, except 11
11 is the perfect/prime task.  I believe it was to look for perfect
number by finding the even prime numbers(?).  I've not written this, as
I don't fully get the math recursion and don't want to just copy/paste a
formula.

Other fun changes were to pass a function as a parameter for the button
callback.  Added a time elapsed when creating the gcd, which I did in
two different ways (binary and euclid)